refactor: move key_length to separate mod
[ci skip]
